refactor: improve navbar rendering
